fix for "error when uploading group image"

+++ class/common/Resizer.php
@@ -79,14 +79,14 @@ class Resizer
                 } else {
                     $new_width  = $this->maxWidth;
                     $divisor    = $width / $new_width;
-                    $new_height = floor($height / $divisor);
+                    $new_height = (int)floor($height / $divisor);
                 }
             } elseif ($height < $this->maxHeight) {
-                $new_height = $height;
+                $new_height = (int)$height;
             } else {
                 $new_height = $this->maxHeight;
                 $divisor    = $height / $new_height;
-                $new_width  = floor($width / $divisor);
+                $new_width  = (int)floor($width / $divisor);
             }
 
             // Create a new temporary image.
